// frontend/src/app/core/api/model/api-response-class-management-stats-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ErrorDetail } from './error-detail';
import { ClassManagementStatsResponse } from './class-management-stats-response';


export interface ApiResponseClassManagementStatsResponse { 
    success?: boolean;
    message?: string;
    errorCode?: string;
    data?: ClassManagementStatsResponse;
    errors?: Array<ErrorDetail>;
    timestamp?: string;
}