// frontend/src/app/core/api/model/class-group-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ClassGroupResponse { 
    id?: number;
    code?: string;
    name?: string;
    capacity?: number;
    currentSize?: number;
    status?: ClassGroupResponse.StatusEnum;
    promotionId?: number;
    promotionName?: string;
    promotionType?: string;
}
export namespace ClassGroupResponse {
    export const StatusEnum = {
        Draft: 'DRAFT',
        Active: 'ACTIVE',
        Inactive: 'INACTIVE',
        Archived: 'ARCHIVED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}